+++ CHANGELOG.md
@@ -9,9 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 ### Changed
+
+- Better support for colour themes and high contrast options
+
 ### Deprecated
+
+- Removed `tna-template--light-theme` option - `tna-template` is light by default
+- Removed `accent-background` and `dark` mixins from colour tools
+
 ### Removed
 ### Fixed
+
+- Fix incorrect `tna-template--black-accent` class
+- Removed all redundant high contrast classes
+
 ### Security
 
 ## [0.1.51](https://github.com/nationalarchives/tna-frontend/compare/v0.1.50...v0.1.51) - 2024-04-04